Considera email na hora de criar

+++ app/services/user_by_csv_creator.rb
@@ -39,29 +39,31 @@ class UserByCsvCreator
   def create_users(entity)
     ActiveRecord::Base.transaction do
       CSV.foreach(file, col_sep: ',', skip_blanks: true) do |new_user|
-        User.find_or_initialize_by(login: new_user[3]).tap do |user|
-          if new_user[5] == '0'
-            user.destroy
-            next
-          end
-          password = new_user[4] || SecureRandom.hex(8)
-          user.login = new_user[3]
-          user.email = new_user[2]
+        unless User.exists?(email: new_user[2])
+          User.find_or_initialize_by(login: new_user[3]).tap do |user|
+            if new_user[5] == '0'
+              user.destroy
+              next
+            end
+            password = new_user[4] || SecureRandom.hex(8)
+            user.login = new_user[3]
+            user.email = new_user[2]
 
-          user.password = password
-          user.password_confirmation = password
+            user.password = password
+            user.password_confirmation = password
 
-          user.status = 'active'
-          user.kind = 'employee'
-          user.admin = true
-          user.receive_news = false
-          user.first_name = new_user[0]
-          user.last_name = new_user[1]
+            user.status = 'active'
+            user.kind = 'employee'
+            user.admin = true
+            user.receive_news = false
+            user.first_name = new_user[0]
+            user.last_name = new_user[1]
 
-          user.save!
+            user.save!
 
-          if set_admin_role(user) && send_mail && new_user[4].nil?
-            UserMailer.delay.by_csv(user.login, user.first_name, user.email, password, entity.domain)
+            if set_admin_role(user) && send_mail && new_user[4].nil?
+              UserMailer.delay.by_csv(user.login, user.first_name, user.email, password, entity.domain)
+            end
           end
         end
       end
